requirements.txt
torch>=2.0.0
transformers>=4.30.0
aiohttp>=3.8.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0  # For async test support
openai>=1.0.0  # For both O3 and Perplexity API calls
python-whois>=0.8.0  # For domain registration checks
dnspython>=2.4.0  # For DNS record checks
requests>=2.31.0  # For HTTP requests
cryptography>=41.0.0  # For SSL certificate handling
tenacity>=8.2.0  # For robust retry logic
tldextract>=5.1.1  # For domain extraction and analysis
langchain>=0.1.0  # For RAG implementation
faiss-cpu>=1.7.4  # For vector storage
tiktoken>=0.5.1  # For token counting
pinecone-client>=3.0.0  # For vector similarity search 